# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POPIA Guard",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Creates a new account. Password is hashed with bcrypt before\nstorage; 409 if the email is already registered.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Exchanges email + password for a JWT bearer token. The token\ncarries no role/permission claims \u2014 see docs/DATABASE.md on why\nRBAC is out of scope for this project.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List Scans",
        "description": "Lists the authenticated user's scans, most recent first. Does not\ninclude findings or score \u2014 use GET /scans/{id} for the full result.",
        "operationId": "list_scans_api_v1_scans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScanJobRead"
                  },
                  "type": "array",
                  "title": "Response List Scans Api V1 Scans Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create Scan",
        "description": "Uploads a .zip archive, scans it synchronously, and returns the\nfindings, risk score, and compliance percentage. Also triggers a\nbest-effort report upload to S3 (see docs/AWS_INTEGRATION.md) \u2014 a\nstorage failure doesn't fail this request.",
        "operationId": "create_scan_api_v1_scans_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_scan_api_v1_scans_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResultRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/scans/{scan_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan",
        "description": "Returns the scan job plus its findings and a freshly-computed score.\n\nThe score isn't persisted on the ScanJob itself \u2014 it's recomputed\nfrom the persisted findings on every request. Cheap at this data\nsize, and keeps scoring logic in exactly one place rather than\nduplicating it in the frontend. (A persisted, versioned Report\nrow does exist separately \u2014 see GET /scans/{id}/report \u2014 but it\nstores the score as of when the report was generated, not a live\nrecomputation.)",
        "operationId": "get_scan_api_v1_scans__scan_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResultRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/findings": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan Findings",
        "description": "Returns just the findings for a scan, without the job metadata or\nscore \u2014 a lighter-weight alternative to GET /scans/{id} for callers\nthat only need the findings list.",
        "operationId": "get_scan_findings_api_v1_scans__scan_id__findings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingRead"
                  },
                  "title": "Response Get Scan Findings Api V1 Scans  Scan Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/report": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan Report",
        "description": "Returns a fresh presigned URL for downloading the report.\n\n404 covers two distinct cases the caller can't tell apart from the\nresponse alone: the scan doesn't exist/isn't yours, or the report\nwas never stored (e.g. S3 was unreachable when the scan completed \u2014\nsee scan_service._try_store_report). Both mean \"nothing to download\nright now,\" which is the only thing this endpoint promises.",
        "operationId": "get_scan_report_api_v1_scans__scan_id__report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportUrlRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "description": "Liveness check used by Docker Compose and, later, CI.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_scan_api_v1_scans_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_scan_api_v1_scans_post"
      },
      "FindingRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "line_number": {
            "type": "integer",
            "title": "Line Number"
          },
          "matched_snippet": {
            "type": "string",
            "title": "Matched Snippet"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "rule_id",
          "severity",
          "file_path",
          "line_number",
          "matched_snippet",
          "created_at"
        ],
        "title": "FindingRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "ReportUrlRead": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_in"
        ],
        "title": "ReportUrlRead"
      },
      "ScanJobRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_reference": {
            "type": "string",
            "title": "Source Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "files_scanned": {
            "type": "integer",
            "title": "Files Scanned"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type",
          "source_reference",
          "status",
          "files_scanned",
          "started_at",
          "completed_at",
          "created_at"
        ],
        "title": "ScanJobRead"
      },
      "ScanResultRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_reference": {
            "type": "string",
            "title": "Source Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "files_scanned": {
            "type": "integer",
            "title": "Files Scanned"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingRead"
            },
            "type": "array",
            "title": "Findings"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "compliance_percentage": {
            "type": "number",
            "title": "Compliance Percentage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type",
          "source_reference",
          "status",
          "files_scanned",
          "started_at",
          "completed_at",
          "created_at",
          "findings",
          "risk_score",
          "compliance_percentage"
        ],
        "title": "ScanResultRead",
        "description": "Returned immediately after a scan completes: the job plus its\nfindings and computed score. Not persisted as a `Report` row yet \u2014\nreport persistence + S3 upload land in Phase 4."
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "UserCreate"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "is_active",
          "created_at"
        ],
        "title": "UserRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}